Refactored error handlers into blueprint
