+++ app/errors/handlers.py
@@ -1,0 +1,27 @@
+# -*- coding: utf-8 -*-
+"""
+    errors
+    ~~~~~~
+
+    Error handling for the application.
+"""
+
+from flask import render_template
+from app import db
+from app.errors import bp
+
+
+@bp.app_errorhandler(403)
+def forbidden_error(error):
+    return render_template('errors/403.html'), 403
+
+
+@bp.app_errorhandler(404)
+def not_found_error(error):
+    return render_template('errors/404.html'), 404
+
+
+@bp.app_errorhandler(500)
+def internal_error(error):
+    db.session.rollback()
+    return render_template('errors/500.html'), 500
